stanza function and test update
stanza function for tokens and a function for sentence segemntation as well as test in PyTest for both functions
