+++ tests/stanza_test.py
@@ -1,7 +1,7 @@
 
 import pytest
 
-from pynlpl.local.text.text_stanza import Stanza_English_default
+from pynlpl.local.text.stanza import Stanza_English_default
 
 lang_model = Stanza_English_default()
 
@@ -10,3 +10,6 @@ text_for_testing = "PyNLP is the universal open source NLP library for Python. I
 
 def test_stanza_sentence_segementation():
   assert lang_model.sentence_segmentation(text_for_testing) == ["PyNLP is the universal open source NLP library for Python.", "It is maintained by The Text API team."]
+
+def test_token_segmentation():
+  assert lang_model.token_segmentation(text_for_testing) == ['PyNLP', 'is', 'the', 'universal', 'open', 'source', 'NLP', 'library', 'for', 'Python', '.', 'It', 'is', 'maintained', 'by', 'The', 'Text', 'API', 'team', '.']
